requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.5
aiosqlite
SQLAlchemy>=2.0.20
aiofiles>=23.1.0
qrcode>=7.4.2
Pillow>=10.0.1